remove mock/spectra and mock/selection
