+++ py/desitarget/mock/quicksurvey.py
@@ -1995,3 +1995,1262 @@ def read_lya(mock_dir_name, target_name='QSO', rand=None, bricksize=0.25,
             'TRUESPECTYPE': 'QSO', 'TEMPLATETYPE': 'QSO', 'TEMPLATESUBTYPE': 'LYA',
             'MAG': mag_g, 'FILES': ofile_list, 'N_PER_FILE': n_per_file}
 
+
+#############
+# Code previously in desitarget.mock.spectra
+#############
+
+import numpy as np
+from desisim.io import read_basis_templates, empty_metatable
+
+class TemplateKDTree(object):
+    """Build a KD Tree for each object type.
+
+    """
+    def __init__(self, nproc=1, verbose=False):
+        from speclite import filters
+        from scipy.spatial import cKDTree as KDTree
+        from desiutil.log import get_logger, DEBUG
+
+        if verbose:
+            self.log = get_logger(DEBUG)
+        else:
+            self.log = get_logger()
+            
+        self.nproc = nproc
+        self.verbose = verbose
+
+        self.bgs_meta = read_basis_templates(objtype='BGS', onlymeta=True)#, verbose=False)
+        self.elg_meta = read_basis_templates(objtype='ELG', onlymeta=True)#, verbose=False)
+        self.lrg_meta = read_basis_templates(objtype='LRG', onlymeta=True)#, verbose=False)
+        self.qso_meta = read_basis_templates(objtype='QSO', onlymeta=True)#, verbose=False)
+        self.wd_da_meta = read_basis_templates(objtype='WD', subtype='DA', onlymeta=True)#, verbose=False)
+        self.wd_db_meta = read_basis_templates(objtype='WD', subtype='DB', onlymeta=True)#, verbose=False)
+
+        self.decamwise = filters.load_filters('decam2014-g', 'decam2014-r', 'decam2014-z',
+                                              'wise2010-W1', 'wise2010-W2')
+
+        # Read all the stellar spectra and synthesize DECaLS/WISE fluxes.
+        self.star_phot()
+
+        self.bgs_tree = KDTree(self._bgs())
+        self.elg_tree = KDTree(self._elg())
+        #self.lrg_tree = KDTree(self._lrg())
+        #self.qso_tree = KDTree(self._qso())
+        self.star_tree = KDTree(self._star())
+        self.wd_da_tree = KDTree(self._wd_da())
+        self.wd_db_tree = KDTree(self._wd_db())
+
+    def star_phot(self, normfilter='decam2014-r'):
+        """Synthesize photometry for the full set of stellar templates."""
+
+        star_flux, star_wave, star_meta = read_basis_templates(objtype='STAR')#, verbose=False)
+        star_maggies_table = self.decamwise.get_ab_maggies(star_flux, star_wave, mask_invalid=True)
+
+        star_maggies = dict()
+        for key in star_maggies_table.columns:
+            star_maggies[key] = star_maggies_table[key] / star_maggies_table[normfilter] # normalized maggies
+        self.star_flux_g = star_maggies['decam2014-g']
+        self.star_flux_r = star_maggies['decam2014-r']
+        self.star_flux_z = star_maggies['decam2014-z']
+        self.star_flux_w1 = star_maggies['wise2010-W1']
+        self.star_flux_w2 = star_maggies['wise2010-W1']
+        
+        self.star_meta = star_meta
+
+    def _bgs(self):
+        """Quantities we care about: redshift (z), M_0.1r, and 0.1(g-r).  This needs to
+        be generalized to accommodate other mocks!
+
+        """
+        zobj = self.bgs_meta['Z'].data
+        mabs = self.bgs_meta['SDSS_UGRIZ_ABSMAG_Z01'].data
+        rmabs = mabs[:, 2]
+        gr = mabs[:, 1] - mabs[:, 2]
+        return np.vstack((zobj, rmabs, gr)).T
+
+    def _elg(self):
+        """Quantities we care about: redshift, g-r, r-z."""
+        
+        zobj = self.elg_meta['Z'].data
+        gr = self.elg_meta['DECAM_G'].data - self.elg_meta['DECAM_R'].data
+        rz = self.elg_meta['DECAM_R'].data - self.elg_meta['DECAM_Z'].data
+        #W1W2 = self.elg_meta['W1'].data - self.elg_meta['W2'].data
+        return np.vstack((zobj, gr, rz)).T
+
+    def _lrg(self):
+        """Quantities we care about: r-z, r-W1."""
+        
+        zobj = self.elg_meta['Z'].data
+        gr = self.elg_meta['DECAM_G'].data - self.elg_meta['DECAM_R'].data
+        rz = self.elg_meta['DECAM_R'].data - self.elg_meta['DECAM_Z'].data
+        #W1W2 = self.elg_meta['W1'].data - self.elg_meta['W2'].data
+        return np.vstack((zobj, gr, rz)).T
+
+    def _star(self):
+        """Quantities we care about: Teff, logg, and [Fe/H].
+
+        """
+        teff = self.star_meta['TEFF'].data
+        logg = self.star_meta['LOGG'].data
+        feh = self.star_meta['FEH'].data
+        return np.vstack((teff, logg, feh)).T
+
+    #def qso(self):
+    #    """Quantities we care about: redshift, XXX"""
+    #    pass 
+
+    def _wd_da(self):
+        """DA white dwarf.  Quantities we care about: Teff and logg.
+
+        """
+        teff = self.wd_da_meta['TEFF'].data
+        logg = self.wd_da_meta['LOGG'].data
+        return np.vstack((teff, logg)).T
+
+    def _wd_db(self):
+        """DB white dwarf.  Quantities we care about: Teff and logg.
+
+        """
+        teff = self.wd_db_meta['TEFF'].data
+        logg = self.wd_db_meta['LOGG'].data
+        return np.vstack((teff, logg)).T
+
+    def query(self, objtype, matrix, subtype=''):
+        """Return the nearest template number based on the KD Tree.
+
+        Args:
+          objtype (str): object type
+          matrix (numpy.ndarray): (M,N) array (M=number of properties,
+            N=number of objects) in the same format as the corresponding
+            function for each object type (e.g., self.bgs).
+          subtype (str, optional): subtype (only for white dwarfs)
+
+        Returns:
+          dist: distance to nearest template
+          indx: index of nearest template
+        
+        """
+        if objtype.upper() == 'BGS':
+            dist, indx = self.bgs_tree.query(matrix)
+            
+        elif objtype.upper() == 'ELG':
+            dist, indx = self.elg_tree.query(matrix)
+            
+        elif objtype.upper() == 'LRG':
+            dist, indx = self.lrg_tree.query(matrix)
+            
+        elif objtype.upper() == 'STAR':
+            dist, indx = self.star_tree.query(matrix)
+            
+        elif objtype.upper() == 'QSO':
+            dist, indx = self.qso_tree.query(matrix)
+            
+        elif objtype.upper() == 'WD':
+            if subtype.upper() == 'DA':
+                dist, indx = self.wd_da_tree.query(matrix)
+            elif subtype.upper() == 'DB':
+                dist, indx = self.wd_db_tree.query(matrix)
+            else:
+                self.log.warning('Unrecognized SUBTYPE {}!'.format(subtype))
+                raise ValueError
+                
+        return dist, indx
+
+class MockSpectra(object):
+    """Generate spectra for each type of mock.  Currently just choose the closest
+    template; we can get fancier later.
+
+    ToDo (@moustakas): apply Galactic extinction.
+
+    """
+    def __init__(self, wavemin=None, wavemax=None, dw=0.2, nproc=1,
+                 rand=None, verbose=False):
+
+        from desimodel.io import load_throughput
+
+        self.tree = TemplateKDTree(nproc=nproc, verbose=verbose)
+
+        # Build a default (buffered) wavelength vector.
+        if wavemin is None:
+            wavemin = load_throughput('b').wavemin - 10.0
+        if wavemax is None:
+            wavemax = load_throughput('z').wavemax + 10.0
+            
+        self.wavemin = wavemin
+        self.wavemax = wavemax
+        self.dw = dw
+        self.wave = np.arange(round(wavemin, 1), wavemax, dw)
+
+        self.rand = rand
+        self.verbose = verbose
+
+        #self.__normfilter = 'decam2014-r' # default normalization filter
+
+        # Initialize the templates once:
+        from desisim.templates import BGS, ELG, LRG, QSO, SIMQSO, STAR, WD
+        self.bgs_templates = BGS(wave=self.wave, normfilter='sdss2010-r') # Need to generalize this!
+        self.elg_templates = ELG(wave=self.wave, normfilter='decam2014-r')
+        self.lrg_templates = LRG(wave=self.wave, normfilter='decam2014-z')
+        self.qso_templates = QSO(wave=self.wave, normfilter='decam2014-g')
+        self.simqso_templates = SIMQSO(wave=self.wave, normfilter='decam2014-g')
+        self.lya_templates = QSO(wave=self.wave, normfilter='decam2014-g')
+        self.star_templates = STAR(wave=self.wave, normfilter='decam2014-r')
+        self.wd_da_templates = WD(wave=self.wave, normfilter='decam2014-g', subtype='DA')
+        self.wd_db_templates = WD(wave=self.wave, normfilter='decam2014-g', subtype='DB')
+        
+    def bgs(self, data, index=None, mockformat='durham_mxxl_hdf5'):
+        """Generate spectra for BGS.
+
+        Currently only the MXXL (durham_mxxl_hdf5) mock is supported.  DATA
+        needs to have Z, SDSS_absmag_r01, SDSS_01gr, VDISP, and SEED, which are
+        assigned in mock.io.read_durham_mxxl_hdf5.  See also
+        TemplateKDTree.bgs().
+
+        """
+        objtype = 'BGS'
+        if index is None:
+            index = np.arange(len(data['Z']))
+            
+        input_meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            input_meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() == 'durham_mxxl_hdf5':
+            alldata = np.vstack((data['Z'][index],
+                                 data['SDSS_absmag_r01'][index],
+                                 data['SDSS_01gr'][index])).T
+            _, templateid = self.tree.query(objtype, alldata)
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+
+        input_meta['TEMPLATEID'] = templateid
+        flux, _, meta = self.bgs_templates.make_templates(input_meta=input_meta,
+                                                          nocolorcuts=True, novdisp=False,
+                                                          verbose=self.verbose)
+
+        return flux, meta
+
+    def elg(self, data, index=None, mockformat='gaussianfield'):
+        """Generate spectra for the ELG sample.
+
+        Currently only the GaussianField mock sample is supported.  DATA needs
+        to have Z, GR, RZ, VDISP, and SEED, which are assigned in
+        mock.io.read_gaussianfield.  See also TemplateKDTree.elg().
+
+        """
+        objtype = 'ELG'
+        if index is None:
+            index = np.arange(len(data['Z']))
+
+        input_meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            input_meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() == 'gaussianfield':
+            alldata = np.vstack((data['Z'][index],
+                                 data['GR'][index],
+                                 data['RZ'][index])).T
+            _, templateid = self.tree.query(objtype, alldata)
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+
+        if False:
+            import matplotlib.pyplot as plt
+            def elg_colorbox(ax):
+                """Draw the ELG selection box."""
+                from matplotlib.patches import Polygon
+                grlim = ax.get_ylim()
+                coeff0, coeff1 = (1.15, -0.15), (-1.2, 1.6)
+                rzmin, rzpivot = 0.3, (coeff1[1] - coeff0[1]) / (coeff0[0] - coeff1[0])
+                verts = [(rzmin, grlim[0]),
+                         (rzmin, np.polyval(coeff0, rzmin)),
+                         (rzpivot, np.polyval(coeff1, rzpivot)),
+                         ((grlim[0] - 0.1 - coeff1[1]) / coeff1[0], grlim[0] - 0.1)
+                         ]
+                ax.add_patch(Polygon(verts, fill=False, ls='--', color='k'))
+
+            fig, ax = plt.subplots()
+            ax.scatter(data['RZ'][index], data['GR'][index])
+            ax.set_xlim(-0.5, 2) ; plt.ylim(-0.5, 2)
+            elg_colorbox(ax)
+            plt.show()
+            import pdb ; pdb.set_trace()
+
+        input_meta['TEMPLATEID'] = templateid
+        flux, _, meta = self.elg_templates.make_templates(input_meta=input_meta,
+                                                          nocolorcuts=True, novdisp=False,
+                                                          verbose=self.verbose)
+
+        return flux, meta
+
+    def elg_test(self, data, index=None, mockformat='gaussianfield'):
+        """Test script -- generate spectra for the ELG sample.
+
+        Currently only the GaussianField mock sample is supported.  DATA needs
+        to have Z, GR, RZ, VDISP, and SEED, which are assigned in
+        mock.io.read_gaussianfield.  See also TemplateKDTree.elg().
+
+        """
+        objtype = 'ELG'
+        if index is None:
+            index = np.arange(len(data['Z']))
+
+        input_meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            input_meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() == 'gaussianfield':
+            alldata = np.vstack((data['Z'][index],
+                                 data['GR'][index],
+                                 data['RZ'][index])).T
+            _, templateid = self.tree.query(objtype, alldata)
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+
+        import matplotlib.pyplot as plt
+        from scipy.interpolate import interp1d
+
+        f1 = interp1d(np.squeeze(self.tree.elg_kcorr['REDSHIFT']), np.squeeze(self.tree.elg_kcorr['GR']), axis=0)
+        gr = f1(data['Z'][index])
+        plt.plot(np.squeeze(self.tree.elg_kcorr['REDSHIFT']), np.squeeze(self.tree.elg_kcorr['GR'])[:, 500])
+        plt.scatter(data['Z'][index], gr[:, 500], marker='x', color='red', s=15)
+        plt.show()
+
+        def elg_colorbox(ax):
+            """Draw the ELG selection box."""
+            from matplotlib.patches import Polygon
+            grlim = ax.get_ylim()
+            coeff0, coeff1 = (1.15, -0.15), (-1.2, 1.6)
+            rzmin, rzpivot = 0.3, (coeff1[1] - coeff0[1]) / (coeff0[0] - coeff1[0])
+            verts = [(rzmin, grlim[0]),
+                     (rzmin, np.polyval(coeff0, rzmin)),
+                     (rzpivot, np.polyval(coeff1, rzpivot)),
+                     ((grlim[0] - 0.1 - coeff1[1]) / coeff1[0], grlim[0] - 0.1)
+                     ]
+            ax.add_patch(Polygon(verts, fill=False, ls='--', color='k'))
+
+        fig, ax = plt.subplots()
+        ax.scatter(data['RZ'][index], data['GR'][index])
+        ax.set_xlim(-0.5, 2) ; plt.ylim(-0.5, 2)
+        elg_colorbox(ax)
+        plt.show()
+        import pdb ; pdb.set_trace()
+
+        input_meta['TEMPLATEID'] = templateid
+        flux, _, meta = self.elg_templates.make_templates(input_meta=input_meta,
+                                                          nocolorcuts=True, novdisp=False,
+                                                          verbose=self.verbose)
+
+        return flux, meta
+
+    def lrg(self, data, index=None, mockformat='gaussianfield'):
+        """Generate spectra for the LRG sample.
+
+        Currently only the GaussianField mock sample is supported.  DATA needs
+        to have Z, GR, RZ, VDISP, and SEED, which are assigned in
+        mock.io.read_gaussianfield.  See also TemplateKDTree.lrg().
+
+        """
+        objtype = 'LRG'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+
+        input_meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            input_meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() == 'gaussianfield':
+            # This is wrong: choose a template with equal probability.
+            templateid = self.rand.choice(self.tree.lrg_meta['TEMPLATEID'], len(index))
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+
+        input_meta['TEMPLATEID'] = templateid
+        flux, _, meta = self.lrg_templates.make_templates(input_meta=input_meta,
+                                                          nocolorcuts=True, novdisp=False,
+                                                          verbose=self.verbose)
+
+        return flux, meta
+
+    def mws(self, data, index=None, mockformat='galaxia'):
+        """Generate spectra for the MWS_NEARBY and MWS_MAIN samples.
+
+        """
+        objtype = 'STAR'
+        if index is None:
+            index = np.arange(len(data['Z']))
+
+        input_meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'TEFF', 'LOGG', 'FEH'),
+                                  ('SEED', 'MAG', 'Z', 'TEFF', 'LOGG', 'FEH')):
+            input_meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() == '100pc':
+            alldata = np.vstack((data['TEFF'][index],
+                                 data['LOGG'][index],
+                                 data['FEH'][index])).T
+            _, templateid = self.tree.query(objtype, alldata)
+            
+        elif mockformat.lower() == 'galaxia':
+            alldata = np.vstack((data['TEFF'][index],
+                                 data['LOGG'][index],
+                                 data['FEH'][index])).T
+            _, templateid = self.tree.query(objtype, alldata)
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+
+        input_meta['TEMPLATEID'] = templateid
+        flux, _, meta = self.star_templates.make_templates(input_meta=input_meta,
+                                                          verbose=self.verbose) # Note! No colorcuts.
+
+        return flux, meta
+
+    def mws_nearby(self, data, index=None, mockformat='100pc'):
+        """Generate spectra for the MWS_NEARBY sample.
+
+        """
+        flux, meta = self.mws(data, index=index, mockformat=mockformat)
+        return flux, meta
+
+    def mws_main(self, data, index=None, mockformat='galaxia'):
+        """Generate spectra for the MWS_MAIN sample.
+
+        """
+        flux, meta = self.mws(data, index=index, mockformat=mockformat)
+        return flux, meta
+
+    def faintstar(self, data, index=None, mockformat='galaxia'):
+        """Generate spectra for the FAINTSTAR (faint stellar) sample.
+
+        """
+        flux, meta = self.mws(data, index=index, mockformat=mockformat)
+        return flux, meta
+
+    def mws_wd(self, data, index=None, mockformat='wd'):
+        """Generate spectra for the MWS_WD sample.  Deal with DA vs DB white dwarfs
+        separately.
+
+        """
+        objtype = 'WD'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+
+        input_meta = empty_metatable(nmodel=nobj, objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'TEFF', 'LOGG', 'SUBTYPE'),
+                                  ('SEED', 'MAG', 'Z', 'TEFF', 'LOGG', 'TEMPLATESUBTYPE')):
+            input_meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() == 'wd':
+            meta = empty_metatable(nmodel=nobj, objtype=objtype)
+            flux = np.zeros([nobj, len(self.wave)], dtype='f4')
+            
+            for subtype in ('DA', 'DB'):
+                these = np.where(input_meta['SUBTYPE'] == subtype)[0]
+                if len(these) > 0:
+                    alldata = np.vstack((data['TEFF'][index][these],
+                                         data['LOGG'][index][these])).T
+                    _, templateid = self.tree.query(objtype, alldata, subtype=subtype)
+
+                    input_meta['TEMPLATEID'][these] = templateid
+                    
+                    template_function = 'wd_{}_templates'.format(subtype.lower())
+                    flux1, _, meta1 = getattr(self, template_function).make_templates(input_meta=input_meta[these],
+                                                          verbose=self.verbose)
+                    
+                    meta[these] = meta1
+                    flux[these, :] = flux1
+            
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+
+        return flux, meta
+
+    def qso(self, data, index=None, mockformat='gaussianfield'):
+        """Generate spectra for the QSO or QSO/LYA samples.
+
+        Note: We need to make sure NORMFILTER matches!
+
+        """
+        from desisim.lya_spectra import get_spectra
+        from desisim.lya_spectra import read_lya_skewers,apply_lya_transmission
+        import fitsio
+
+        objtype = 'QSO'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+
+        if mockformat.lower() == 'gaussianfield':
+            input_meta = empty_metatable(nmodel=nobj, objtype=objtype)
+            for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT'),
+                                      ('SEED', 'MAG', 'Z')):
+                input_meta[inkey] = data[datakey][index]
+
+            # Build the tracer and Lya forest QSO spectra separately.
+            meta = empty_metatable(nmodel=nobj, objtype=objtype)
+            flux = np.zeros([nobj, len(self.wave)], dtype='f4')
+
+            lya = np.where( data['TEMPLATESUBTYPE'][index] == 'LYA' )[0]
+            tracer = np.where( data['TEMPLATESUBTYPE'][index] == '' )[0]
+
+            if len(tracer) > 0:
+                flux1, _, meta1 = self.qso_templates.make_templates(input_meta=input_meta[tracer],
+                                                                    lyaforest=False,
+                                                                    nocolorcuts=True,
+                                                                    verbose=self.verbose)
+                meta[tracer] = meta1
+                flux[tracer, :] = flux1
+
+            if len(lya) > 0:
+                ilya = index[lya].astype(int)
+                nqso = ilya.size
+                                
+                if 'LYAHDU' in data : 
+                    # this is the old format with one HDU per spectrum
+                    alllyafile = data['LYAFILES'][ilya]
+                    alllyahdu = data['LYAHDU'][ilya]
+
+                    for lyafile in sorted(set(alllyafile)):
+                        these = np.where( lyafile == alllyafile )[0]
+
+                        templateid = alllyahdu[these] - 1 # templateid is 0-indexed
+                        flux1, _, meta1 = get_spectra(lyafile, templateid=templateid, normfilter=data['FILTERNAME'],
+                                                      rand=self.rand, qso=self.lya_templates, nocolorcuts=True)
+                        meta1['SUBTYPE'] = 'LYA'
+                        meta[lya[these]] = meta1
+                        flux[lya[these], :] = flux1
+                else : # new format
+                    # Read skewers.
+                    skewer_wave = None
+                    skewer_trans = None
+                    skewer_meta = None
+                    
+                    # All the files that contain at least one QSO skewer.
+                    alllyafile = data['LYAFILES'][ilya]
+                    uniquelyafiles = sorted(set(alllyafile))
+                                        
+                    for lyafile in uniquelyafiles:
+                        these = np.where( alllyafile == lyafile )[0]
+                        objid_in_data = data['OBJID'][ilya][these]
+                        objid_in_mock = (fitsio.read(lyafile, columns=['MOCKID'], upper=True,
+                                                     ext=1).astype(float)).astype(int)
+                        o2i = dict()
+                        for i, o in enumerate(objid_in_mock):
+                            o2i[o] = i
+                        indices_in_mock_healpix = np.zeros(objid_in_data.size).astype(int)
+                        for i, o in enumerate(objid_in_data):
+                            if not o in o2i:
+                                self.log.error("No MOCKID={} in {}. It's a bug, should never happen".format(o,lyafile))
+                                raise(KeyError("No MOCKID={} in {}. It's a bug, should never happen".format(o,lyafile)))
+                            indices_in_mock_healpix[i] = o2i[o]
+                        
+                        tmp_wave, tmp_trans, tmp_meta = read_lya_skewers(lyafile, indices=indices_in_mock_healpix) 
+                                                
+                        if skewer_wave is None:
+                            skewer_wave = tmp_wave
+                            dw = skewer_wave[1]-skewer_wave[0] # this is just to check same wavelength
+                            skewer_trans = np.zeros((nqso,skewer_wave.size)) # allocate skewer_array
+                            skewer_meta = dict()
+                            for k in tmp_meta.dtype.names:
+                                skewer_meta[k] = np.zeros(nqso).astype(tmp_meta[k].dtype)
+                        else :
+                            # check wavelength is the same for all skewers
+                            assert(np.max(np.abs(wave-tmp_wave))<0.001*dw)
+                        
+                        skewer_trans[these] = tmp_trans
+                        for k in skewer_meta.keys():
+                            skewer_meta[k][these] = tmp_meta[k]
+                    
+                    # Check we matched things correctly.
+                    assert(np.max(np.abs(skewer_meta["Z"]-data['Z'][ilya]))<0.000001)
+                    assert(np.max(np.abs(skewer_meta["RA"]-data['RA'][ilya]))<0.000001)
+                    assert(np.max(np.abs(skewer_meta["DEC"]-data['DEC'][ilya]))<0.000001)
+                    
+                    # Now we create a series of QSO spectra all at once, which
+                    # is faster than calling each one at a time.
+                    
+                    #seed = self.rand.randint(2**32)
+                    #qso  = self.lya_templates
+                    qso_flux, qso_wave, qso_meta = self.simqso_templates.make_templates(
+                        nmodel=nqso, redshift=data['Z'][ilya], #seed=seed,
+                        lyaforest=False, nocolorcuts=True)
+                    
+                    # apply transmission to QSOs
+                    qso_flux = apply_lya_transmission(qso_wave, qso_flux, skewer_wave, skewer_trans)
+                    
+                    qso_meta['SUBTYPE'] = 'LYA'
+                    meta[lya] = qso_meta
+                    flux[lya, :] = qso_flux
+                    
+        else:
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            
+        return flux, meta
+
+    def sky(self, data, index=None, mockformat=None):
+        """Generate spectra for SKY.
+
+        """
+        objtype = 'SKY'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+            
+        meta = empty_metatable(nmodel=nobj, objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'REDSHIFT'),
+                                  ('SEED', 'Z')):
+            meta[inkey] = data[datakey][index]
+        flux = np.zeros((nobj, len(self.wave)), dtype='i1')
+
+        return flux, meta
+
+class MockMagnitudes(object):
+    """
+    Generate mock magnitudes for each mock.
+    """
+    def __init__(self, nproc=1, rand=None, verbose=False):
+        self.rand = rand
+        self.verbose = verbose
+    def bgs(self, data, index=None, mockformat='durham_mxxl_hdf5'):
+        """Generate magnitudes for BGS.
+        Currently only the MXXL (durham_mxxl_hdf5) mock is supported.  DATA
+        needs to have Z, SDSS_absmag_r01, SDSS_01gr, VDISP, and SEED, which are
+        assigned in mock.io.read_durham_mxxl_hdf5.  
+        """
+        objtype = 'BGS'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() != 'durham_mxxl_hdf5':
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            return meta
+        meta['FLUX_R'][:] = 10**((22.5 - data['MAG'][index])/2.5) # r-band flux
+        meta['FLUX_G'][:] = 10**((22.5-(data['SDSS_01gr'][index] + data['MAG'][index]))/2.5) # g-band flux
+        return meta
+    
+    def mws(self, data, index=None, mockformat='galaxia'):
+        """Generate magnitudes for the MWS_NEARBY and MWS_MAIN samples.
+        """
+        objtype = 'STAR'
+        if index is None:
+            index = np.arange(len(data['Z']))
+
+        meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'TEFF', 'LOGG', 'FEH'),
+                                  ('SEED', 'MAG', 'Z', 'TEFF', 'LOGG', 'FEH')):
+            meta[inkey] = data[datakey][index]
+
+        if not (mockformat.lower() in ['100pc','galaxia']):
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            return meta
+
+        meta['FLUX_R'][:] = 10**((22.5 - data['MAG'][index])/2.5) # r-band flux
+        return meta
+                
+    def mws_nearby(self, data, index=None, mockformat='100pc'):
+        """Generate magnitudes for the MWS_NEARBY sample.
+        """
+        meta = self.mws(data, index=index, mockformat=mockformat)
+        return meta
+
+    def mws_main(self, data, index=None, mockformat='galaxia'):
+        """Generate magnitudes for the MWS_MAIN sample.
+        """
+        meta = self.mws(data, index=index, mockformat=mockformat)
+        return meta
+
+    def faintstar(self, data, index=None, mockformat='galaxia'):
+        """Generate magnitudes for the FAINTSTAR (faint stellar) sample.
+        """
+        meta = self.mws(data, index=index, mockformat=mockformat)
+        return meta
+        
+    def elg(self, data, index=None, mockformat='gaussianfield'):
+        """Generate magnitudes for the ELG sample.
+        Currently only the GaussianField mock sample is supported.  DATA needs
+        to have Z, GR, RZ, VDISP, and SEED, which are assigned in
+        mock.io.read_gaussianfield.  
+        """
+        objtype = 'ELG'
+        if index is None:
+            index = np.arange(len(data['Z']))
+
+        meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() != 'gaussianfield':
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            return meta
+        
+        meta['FLUX_G'][:] = 10**((22.5 - (data['GR'][index] + data['MAG'][index]))/2.5) # g-band flux
+        meta['FLUX_R'][:] = 10**((22.5 - data['MAG'][index])/2.5) # r-band flux
+        meta['FLUX_Z'][:] = 10**((22.5 - (data['MAG'][index] - data['RZ'][index]))/2.5) # z-band flux
+        return meta
+    
+    def lrg(self, data, index=None, mockformat='gaussianfield'):
+        """Generate magnitudes for the LRG sample.
+        Currently only the GaussianField mock sample is supported.  DATA needs
+        to have Z, GR, RZ, VDISP, and SEED, which are assigned in
+        mock.io.read_gaussianfield.  
+        """
+        objtype = 'LRG'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+
+        meta = empty_metatable(nmodel=len(index), objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'VDISP'),
+                                  ('SEED', 'MAG', 'Z', 'VDISP')):
+            meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() != 'gaussianfield':
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            return meta
+        
+        meta['FLUX_G'][:] = 10**((22.5 - (data['GR'][index] + data['RZ'][index] + data['MAG'][index]))/2.5) # g-band flux  
+        meta['FLUX_R'][:] =  10**((22.5 - (data['MAG'][index] + data['RZ'][index]))/2.5) # r-band flux
+        meta['FLUX_Z'][:] = 10**((22.5 - data['MAG'][index])/2.5) # z-band flux
+        meta['FLUX_W1'][:] = 10**((22.5 - (data['RZ'][index] + data['MAG'][index] - data['RW1'][index]))/2.5)# wise flux 1
+        meta['FLUX_W2'][:] = 10**((22.5 - (data['RZ'][index] + data['MAG'][index] - data['RW1'][index] - data['W1W2'][index]))/2.5)# wise flux 2
+        
+        return meta
+    
+    def qso(self, data, index=None, mockformat='gaussianfield'):
+        """Generate magnitudes for the QSO or QSO/LYA samples.
+        """
+        from desisim.lya_spectra import get_spectra
+                
+        objtype = 'QSO'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+        meta = empty_metatable(nmodel=nobj, objtype=objtype)
+
+        if mockformat.lower() != 'gaussianfield':
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            return meta
+        
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT'),
+                                      ('SEED', 'MAG', 'Z')):
+            meta[inkey] = data[datakey][index]
+        
+        meta['FLUX_G'][:] = 10**((22.5 - data['MAG'][index])/2.5) # g-band flux
+        meta['FLUX_R'][:] = 10**((22.5 - (data['MAG'][index] - data['GR'][index]))/2.5) # r-band flux
+        meta['FLUX_Z'][:] = 10**((22.5 - (data['MAG'][index] - data['GR'][index] - data['RZ'][index]))/2.5) # z-band flux
+        meta['FLUX_W1'][:] = 10**((22.5 - (data['MAG'][index] - data['GR'][index] - data['RW1'][index]))/2.5)# wise flux 1
+        meta['FLUX_W2'][:] = 10**((22.5 - (data['MAG'][index] - data['GR'][index] - data['RW1'][index] - data['W1W2'][index]))/2.5)# wise flux 2
+        
+        lya = np.where( data['TEMPLATESUBTYPE'][index] == 'LYA' )[0]                 
+        if len(lya) > 0:
+            meta['SUBTYPE'][lya] = 'LYA'
+                                       
+        return meta
+    def mws_wd(self, data, index=None, mockformat='wd'):
+        """Generate magnitudes for the MWS_WD sample.  Deal with DA vs DB white dwarfs
+        separately.
+        """
+        objtype = 'WD'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+
+        meta = empty_metatable(nmodel=nobj, objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'MAG', 'REDSHIFT', 'TEFF', 'LOGG', 'SUBTYPE'),
+                                  ('SEED', 'MAG', 'Z', 'TEFF', 'LOGG', 'TEMPLATESUBTYPE')):
+            meta[inkey] = data[datakey][index]
+
+        if mockformat.lower() != 'wd':
+            raise ValueError('Unrecognized mockformat {}!'.format(mockformat))
+            return meta
+            
+        for subtype in ('DA', 'DB'):
+            these = np.where(meta['SUBTYPE'] == subtype)[0]
+            if len(these) > 0:
+                meta['SUBTYPE'][these] = subtype    
+                meta['TEMPLATEID'][:] = -1
+    
+        meta['FLUX_G'][:] = 10**((22.5 - data['MAG'][index])/2.5) # g-band flux
+        return meta
+    
+    def sky(self, data, index=None, mockformat=None):
+        """Generate spectra for SKY.
+
+        """
+        objtype = 'SKY'
+        if index is None:
+            index = np.arange(len(data['Z']))
+        nobj = len(index)
+            
+        meta = empty_metatable(nmodel=nobj, objtype=objtype)
+        for inkey, datakey in zip(('SEED', 'REDSHIFT'),
+                                  ('SEED', 'Z')):
+            meta[inkey] = data[datakey][index]
+            
+        return meta
+
+
+#############
+# Code previously in desitarget.mock.selection
+#############
+
+import numpy as np
+
+class SelectTargets(object):
+    """Select various types of targets.  Most of this functionality is taken from
+    desitarget.cuts but that code has not been factored in a way that is
+    convenient at this time.
+
+    """
+    def __init__(self, rand=None, brick_info=None, verbose=False):
+        
+        from desitarget import (desi_mask, bgs_mask, mws_mask,
+                                contam_mask, obsconditions)
+        from desiutil.log import get_logger, DEBUG
+
+        if verbose:
+            self.log = get_logger(DEBUG)
+        else:
+            self.log = get_logger()
+        
+        self.desi_mask = desi_mask
+        self.bgs_mask = bgs_mask
+        self.mws_mask = mws_mask
+        self.contam_mask = contam_mask
+        self.obsconditions = obsconditions
+        
+        self.rand = rand
+        self.brick_info = brick_info
+
+    def _star_select(self, targets, truth):
+        """Select stellar (faint and bright) contaminants for the extragalactic
+        targets.
+
+        """ 
+        from desitarget.cuts import isBGS_faint, isELG, isLRG_colors, isQSO_colors
+
+        gflux, rflux, zflux, w1flux, w2flux = targets['FLUX_G'], targets['FLUX_R'], \
+          targets['FLUX_Z'], targets['FLUX_W1'], targets['FLUX_W2']
+
+        # Select stellar contaminants for BGS_FAINT targets.
+        bgs_faint = isBGS_faint(rflux=rflux)
+        targets['BGS_TARGET'] |= (bgs_faint != 0) * self.bgs_mask.BGS_FAINT
+        targets['BGS_TARGET'] |= (bgs_faint != 0) * self.bgs_mask.BGS_FAINT_SOUTH
+        targets['DESI_TARGET'] |= (bgs_faint != 0) * self.desi_mask.BGS_ANY
+        
+        targets['OBSCONDITIONS'] |= (bgs_faint != 0) * self.obsconditions.mask(
+            self.bgs_mask.BGS_FAINT.obsconditions)
+        targets['OBSCONDITIONS'] |= (bgs_faint != 0) * self.obsconditions.mask(
+            self.bgs_mask.BGS_FAINT_SOUTH.obsconditions)
+        targets['OBSCONDITIONS'] |= (bgs_faint != 0) * self.obsconditions.mask(
+            self.desi_mask.BGS_ANY.obsconditions)
+        
+        truth['CONTAM_TARGET'] |= (bgs_faint != 0) * self.contam_mask.BGS_IS_STAR
+        truth['CONTAM_TARGET'] |= (bgs_faint != 0) * self.contam_mask.BGS_CONTAM
+
+        # Select stellar contaminants for ELG targets.
+        elg = isELG(gflux=gflux, rflux=rflux, zflux=zflux)
+        targets['DESI_TARGET'] |= (elg != 0) * self.desi_mask.ELG
+        targets['DESI_TARGET'] |= (elg != 0) * self.desi_mask.ELG_SOUTH
+        
+        targets['OBSCONDITIONS'] |= (elg != 0) * self.obsconditions.mask(
+            self.desi_mask.ELG.obsconditions)
+        targets['OBSCONDITIONS'] |= (elg != 0) * self.obsconditions.mask(
+            self.desi_mask.ELG_SOUTH.obsconditions)
+        
+        truth['CONTAM_TARGET'] |= (elg != 0) * self.contam_mask.ELG_IS_STAR
+        truth['CONTAM_TARGET'] |= (elg != 0) * self.contam_mask.ELG_CONTAM
+
+        # Select stellar contaminants for LRG targets.
+        lrg = isLRG_colors(gflux=gflux, rflux=rflux, zflux=zflux,
+                           w1flux=w1flux, w2flux=w2flux)
+        targets['DESI_TARGET'] |= (lrg != 0) * self.desi_mask.LRG
+        targets['DESI_TARGET'] |= (lrg != 0) * self.desi_mask.LRG_SOUTH
+
+        targets['OBSCONDITIONS'] |= (lrg != 0) * self.obsconditions.mask(
+            self.desi_mask.LRG.obsconditions)
+        targets['OBSCONDITIONS'] |= (lrg != 0) * self.obsconditions.mask(
+            self.desi_mask.LRG_SOUTH.obsconditions)
+        
+        truth['CONTAM_TARGET'] |= (lrg != 0) * self.contam_mask.LRG_IS_STAR
+        truth['CONTAM_TARGET'] |= (lrg != 0) * self.contam_mask.LRG_CONTAM
+
+        # Select stellar contaminants for QSO targets.
+        qso = isQSO_colors(gflux=gflux, rflux=rflux, zflux=zflux, w1flux=w1flux, 
+                           w2flux=w2flux, optical=False) # Note optical=False!
+        targets['DESI_TARGET'] |= (qso != 0) * self.desi_mask.QSO
+        targets['DESI_TARGET'] |= (qso != 0) * self.desi_mask.QSO_SOUTH
+
+        targets['OBSCONDITIONS'] |= (qso != 0) * self.obsconditions.mask(
+            self.desi_mask.QSO.obsconditions)
+        targets['OBSCONDITIONS'] |= (qso != 0) * self.obsconditions.mask(
+            self.desi_mask.QSO_SOUTH.obsconditions)
+        
+        truth['CONTAM_TARGET'] |= (qso != 0) * self.contam_mask.QSO_IS_STAR
+        truth['CONTAM_TARGET'] |= (qso != 0) * self.contam_mask.QSO_CONTAM
+
+    def _std_select(self, targets, truth, boss_std=None):
+        """Select bright- and dark-time standard stars."""
+        from desitarget.cuts import isFSTD
+
+        gflux, rflux, zflux, w1flux, w2flux = targets['FLUX_G'], targets['FLUX_R'], \
+          targets['FLUX_Z'], targets['FLUX_W1'], targets['FLUX_W2']
+
+        obs_rflux = rflux / targets['MW_TRANSMISSION_R'] # attenuate for Galactic dust
+
+        gsnr, rsnr, zsnr = gflux*0+100, rflux*0+100, zflux*0+100    # Hack -- fixed S/N
+        gfracflux, rfracflux, zfracflux = gflux*0, rflux*0, zflux*0 # # No contamination from neighbors.
+        objtype = np.repeat('PSF', len(targets)).astype('U3') # Right data type?!?
+
+        # Select dark-time FSTD targets.  Temporary hack to use the BOSS
+        # standard-star selection algorith.
+        if boss_std is None:
+            fstd = isFSTD(gflux=gflux, rflux=rflux, zflux=zflux, objtype=objtype,
+                          gsnr=gsnr, rsnr=rsnr, zsnr=zsnr, 
+                          gfracflux=gfracflux, rfracflux=rfracflux, zfracflux=zfracflux, 
+                          obs_rflux=obs_rflux)
+        else:
+            rbright, rfaint = 16, 19
+            fstd = boss_std * ( obs_rflux < 10**((22.5 - rbright)/2.5) ) * ( obs_rflux > 10**((22.5 - rfaint)/2.5) )
+
+        targets['DESI_TARGET'] |= (fstd != 0) * self.desi_mask.STD_FSTAR
+        targets['OBSCONDITIONS'] |= (fstd != 0) * self.obsconditions.mask(
+            self.desi_mask.STD_FSTAR.obsconditions)
+
+        # Select bright-time FSTD targets.  Temporary hack to use the BOSS
+        # standard-star selection algorith.
+        if boss_std is None:
+            fstd_bright = isFSTD(gflux=gflux, rflux=rflux, zflux=zflux, objtype=objtype,
+                                 gsnr=gsnr, rsnr=rsnr, zsnr=zsnr, 
+                                 gfracflux=gfracflux, rfracflux=rfracflux, zfracflux=zfracflux, 
+                                 obs_rflux=obs_rflux, bright=True)
+        else:
+            rbright, rfaint = 14, 17
+            fstd_bright = boss_std * ( obs_rflux < 10**((22.5 - rbright)/2.5) ) * ( obs_rflux > 10**((22.5 - rfaint)/2.5) )
+
+        targets['DESI_TARGET'] |= (fstd_bright != 0) * self.desi_mask.STD_BRIGHT
+        targets['OBSCONDITIONS'] |= (fstd_bright != 0) * self.obsconditions.mask(
+            self.desi_mask.STD_BRIGHT.obsconditions)
+        
+    def bgs_select(self, targets, truth, boss_std=None):
+        """Select BGS targets.
+
+        """
+        from desitarget.cuts import isBGS_bright, isBGS_faint
+
+        rflux = targets['FLUX_R']
+
+        # Select BGS_BRIGHT targets.
+        bgs_bright = isBGS_bright(rflux=rflux)
+        targets['BGS_TARGET'] |= (bgs_bright != 0) * self.bgs_mask.BGS_BRIGHT
+        targets['BGS_TARGET'] |= (bgs_bright != 0) * self.bgs_mask.BGS_BRIGHT_SOUTH
+        targets['DESI_TARGET'] |= (bgs_bright != 0) * self.desi_mask.BGS_ANY
+
+        targets['OBSCONDITIONS'] |= (bgs_bright != 0) * self.obsconditions.mask(
+            self.bgs_mask.BGS_BRIGHT.obsconditions)
+        targets['OBSCONDITIONS'] |= (bgs_bright != 0) * self.obsconditions.mask(
+            self.bgs_mask.BGS_BRIGHT_SOUTH.obsconditions)
+        targets['OBSCONDITIONS'] |= (bgs_bright != 0) * self.obsconditions.mask(
+            self.desi_mask.BGS_ANY.obsconditions)
+        
+        # Select BGS_FAINT targets.
+        bgs_faint = isBGS_faint(rflux=rflux)
+        targets['BGS_TARGET'] |= (bgs_faint != 0) * self.bgs_mask.BGS_FAINT
+        targets['BGS_TARGET'] |= (bgs_faint != 0) * self.bgs_mask.BGS_FAINT_SOUTH
+        targets['DESI_TARGET'] |= (bgs_faint != 0) * self.desi_mask.BGS_ANY
+        
+        targets['OBSCONDITIONS'] |= (bgs_faint != 0) * self.obsconditions.mask(
+            self.bgs_mask.BGS_FAINT.obsconditions)
+        targets['OBSCONDITIONS'] |= (bgs_faint != 0) * self.obsconditions.mask(
+            self.bgs_mask.BGS_FAINT_SOUTH.obsconditions)
+        targets['OBSCONDITIONS'] |= (bgs_faint != 0) * self.obsconditions.mask(
+            self.desi_mask.BGS_ANY.obsconditions)
+
+    def elg_select(self, targets, truth, boss_std=None):
+        """Select ELG targets and contaminants."""
+        from desitarget.cuts import isELG, isQSO_colors
+
+        gflux, rflux, zflux, w1flux, w2flux = targets['FLUX_G'], targets['FLUX_R'], \
+          targets['FLUX_Z'], targets['FLUX_W1'], targets['FLUX_W2']
+        
+        elg = isELG(gflux=gflux, rflux=rflux, zflux=zflux)
+
+        targets['DESI_TARGET'] |= (elg != 0) * self.desi_mask.ELG
+        targets['DESI_TARGET'] |= (elg != 0) * self.desi_mask.ELG_SOUTH
+        targets['OBSCONDITIONS'] |= (elg != 0) * self.obsconditions.mask(
+            self.desi_mask.ELG.obsconditions)
+        targets['OBSCONDITIONS'] |= (elg != 0) * self.obsconditions.mask(
+            self.desi_mask.ELG_SOUTH.obsconditions)
+
+        # Select ELG contaminants for QSO targets.  There should be a morphology
+        # cut here, too, so we're going to overestimate the number of
+        # contaminants.  To make sure we don't reduce the number density of true
+        # ELGs, demand that the QSO contaminants are not also selected as ELGs.
+        qso = isQSO_colors(gflux=gflux, rflux=rflux, zflux=zflux, w1flux=w1flux, 
+                           w2flux=w2flux, optical=False) # Note optical=False!
+        targets['DESI_TARGET'] |= (qso != 0) * (elg == 0) * self.desi_mask.QSO
+        targets['DESI_TARGET'] |= (qso != 0) * (elg == 0) * self.desi_mask.QSO_SOUTH
+        targets['OBSCONDITIONS'] |= (qso != 0) * (elg == 0) * self.obsconditions.mask(
+            self.desi_mask.QSO.obsconditions)
+        targets['OBSCONDITIONS'] |= (qso != 0) * (elg == 0) * self.obsconditions.mask(
+            self.desi_mask.QSO_SOUTH.obsconditions)
+
+        truth['CONTAM_TARGET'] |= (qso != 0) * (elg == 0) * self.contam_mask.QSO_IS_ELG
+        truth['CONTAM_TARGET'] |= (qso != 0) * (elg == 0) * self.contam_mask.QSO_IS_GALAXY
+        truth['CONTAM_TARGET'] |= (qso != 0) * (elg == 0) * self.contam_mask.QSO_CONTAM
+
+    def faintstar_select(self, targets, truth, boss_std=None):
+        """Select faint stellar contaminants for the extragalactic targets.""" 
+
+        self._star_select(targets, truth)
+
+    def lrg_select(self, targets, truth, boss_std=None):
+        """Select LRG targets."""
+        from desitarget.cuts import isLRG_colors, isQSO_colors
+
+        gflux, rflux, zflux, w1flux, w2flux = targets['FLUX_G'], targets['FLUX_R'], \
+          targets['FLUX_Z'], targets['FLUX_W1'], targets['FLUX_W2']
+
+        lrg = isLRG_colors(gflux=gflux, rflux=rflux, zflux=zflux,
+                           w1flux=w1flux, w2flux=w2flux)
+
+        targets['DESI_TARGET'] |= (lrg != 0) * self.desi_mask.LRG
+        targets['DESI_TARGET'] |= (lrg != 0) * self.desi_mask.LRG_SOUTH
+        targets['OBSCONDITIONS'] |= (lrg != 0) * self.obsconditions.mask(
+            self.desi_mask.LRG.obsconditions)
+        targets['OBSCONDITIONS'] |= (lrg != 0) * self.obsconditions.mask(
+            self.desi_mask.LRG_SOUTH.obsconditions)
+        
+        # Select LRG contaminants for QSO targets.  There should be a morphology
+        # cut here, too, so we're going to overestimate the number of
+        # contaminants.  To make sure we don't reduce the number density of true
+        # LRGs, demand that the QSO contaminants are not also selected as LRGs.
+        qso = isQSO_colors(gflux=gflux, rflux=rflux, zflux=zflux, w1flux=w1flux, 
+                           w2flux=w2flux, optical=False) # Note optical=False!
+        targets['DESI_TARGET'] |= (qso != 0) * (lrg == 0) * self.desi_mask.QSO
+        targets['DESI_TARGET'] |= (qso != 0) * (lrg == 0) * self.desi_mask.QSO_SOUTH
+        targets['OBSCONDITIONS'] |= (qso != 0) * (lrg == 0)  * self.obsconditions.mask(
+            self.desi_mask.QSO.obsconditions)
+        targets['OBSCONDITIONS'] |= (qso != 0) * (lrg == 0)  * self.obsconditions.mask(
+            self.desi_mask.QSO_SOUTH.obsconditions)
+        
+        truth['CONTAM_TARGET'] |= (qso != 0) * (lrg == 0) * self.contam_mask.QSO_IS_LRG
+        truth['CONTAM_TARGET'] |= (qso != 0) * (lrg == 0) * self.contam_mask.QSO_IS_GALAXY
+        truth['CONTAM_TARGET'] |= (qso != 0) * (lrg == 0) * self.contam_mask.QSO_CONTAM
+
+    def mws_main_select(self, targets, truth, boss_std=None):
+        """Select MWS_MAIN, MWS_MAIN_VERY_FAINT, standard stars, and (bright)
+        contaminants for extragalactic targets.  The selection here eventually
+        will be done with Gaia (I think).
+
+        """
+        def _isMWS_MAIN(rflux):
+            """A function like this should be in desitarget.cuts. Select 15<r<19 stars."""
+            main = rflux > 10**( (22.5 - 19.0) / 2.5 )
+            main &= rflux <= 10**( (22.5 - 15.0) / 2.5 )
+            return main
+
+        def _isMWS_MAIN_VERY_FAINT(rflux):
+            """A function like this should be in desitarget.cuts. Select 19<r<20 filler stars."""
+            faint = rflux > 10**( (22.5 - 20.0) / 2.5 )
+            faint &= rflux <= 10**( (22.5 - 19.0) / 2.5 )
+            return faint
+        
+        gflux, rflux, zflux, w1flux, w2flux = targets['FLUX_G'], targets['FLUX_R'], \
+          targets['FLUX_Z'], targets['FLUX_W1'], targets['FLUX_W2']
+
+        # Select MWS_MAIN targets.
+        mws_main = _isMWS_MAIN(rflux=rflux)
+        #mws_main = np.ones(len(targets)) # select everything!
+        
+        targets['MWS_TARGET'] |= (mws_main != 0) * self.mws_mask.mask('MWS_MAIN')
+        targets['DESI_TARGET'] |= (mws_main != 0) * self.desi_mask.MWS_ANY
+        targets['OBSCONDITIONS'] |= (mws_main != 0)  * self.obsconditions.mask(
+            self.mws_mask.MWS_MAIN.obsconditions)
+        targets['OBSCONDITIONS'] |= (mws_main != 0)  * self.obsconditions.mask(
+            self.desi_mask.MWS_ANY.obsconditions)
+        
+        mws_main_very_faint = _isMWS_MAIN_VERY_FAINT(rflux=rflux)
+        targets['MWS_TARGET'] |= (mws_main_very_faint != 0) * self.mws_mask.mask('MWS_MAIN_VERY_FAINT')
+        targets['DESI_TARGET'] |= (mws_main_very_faint != 0) * self.desi_mask.MWS_ANY
+        targets['OBSCONDITIONS'] |= (mws_main_very_faint != 0)  * self.obsconditions.mask(
+            self.mws_mask.MWS_MAIN_VERY_FAINT.obsconditions)
+        targets['OBSCONDITIONS'] |= (mws_main_very_faint != 0)  * self.obsconditions.mask(
+            self.desi_mask.MWS_ANY.obsconditions)
+
+        # Select standard stars.
+        self._std_select(targets, truth, boss_std=boss_std)
+        
+        # Select bright stellar contaminants for the extragalactic targets.
+        self._star_select(targets, truth)
+
+    def mws_nearby_select(self, targets, truth, boss_std=None):
+        """Select MWS_NEARBY targets.  The selection eventually will be done with Gaia,
+        so for now just do a "perfect" selection.
+
+        """
+        mws_nearby = np.ones(len(targets)) # select everything!
+        #mws_nearby = (truth['MAG'] <= 20.0) * 1 # SDSS g-band!
+
+        targets['MWS_TARGET'] |= (mws_nearby != 0) * self.mws_mask.mask('MWS_NEARBY')
+        targets['DESI_TARGET'] |= (mws_nearby != 0) * self.desi_mask.MWS_ANY
+        targets['OBSCONDITIONS'] |= (mws_nearby != 0)  * self.obsconditions.mask(
+            self.mws_mask.MWS_NEARBY.obsconditions)
+        targets['OBSCONDITIONS'] |= (mws_nearby != 0)  * self.obsconditions.mask(
+            self.desi_mask.MWS_ANY.obsconditions)
+
+    def mws_wd_select(self, targets, truth, boss_std=None):
+        """Select MWS_WD and STD_WD targets.  The selection eventually will be done with
+        Gaia, so for now just do a "perfect" selection here.
+
+        """
+        #mws_wd = np.ones(len(targets)) # select everything!
+        mws_wd = ((truth['MAG'] >= 15.0) * (truth['MAG'] <= 20.0)) * 1 # SDSS g-band!
+
+        targets['MWS_TARGET'] |= (mws_wd != 0) * self.mws_mask.mask('MWS_WD')
+        targets['DESI_TARGET'] |= (mws_wd != 0) * self.desi_mask.MWS_ANY
+        targets['OBSCONDITIONS'] |= (mws_wd != 0)  * self.obsconditions.mask(
+            self.mws_mask.MWS_WD.obsconditions)
+        targets['OBSCONDITIONS'] |= (mws_wd != 0)  * self.obsconditions.mask(
+            self.desi_mask.MWS_ANY.obsconditions)
+
+        # Select STD_WD; cut just on g-band magnitude (not TEMPLATESUBTYPE!)
+        std_wd = (truth['MAG'] <= 19.0) * 1 # SDSS g-band!
+        targets['DESI_TARGET'] |= (std_wd !=0) * self.desi_mask.mask('STD_WD')
+        targets['OBSCONDITIONS'] |= (std_wd != 0)  * self.obsconditions.mask(
+            self.desi_mask.STD_WD.obsconditions)
+
+    def qso_select(self, targets, truth, boss_std=None):
+        """Select QSO targets and contaminants."""
+        from desitarget.cuts import isQSO_colors, isELG
+
+        gflux, rflux, zflux, w1flux, w2flux = targets['FLUX_G'], targets['FLUX_R'], \
+          targets['FLUX_Z'], targets['FLUX_W1'], targets['FLUX_W2']
+          
+        qso = isQSO_colors(gflux=gflux, rflux=rflux, zflux=zflux,
+                           w1flux=w1flux, w2flux=w2flux, optical=True)
+
+        targets['DESI_TARGET'] |= (qso != 0) * self.desi_mask.QSO
+        targets['DESI_TARGET'] |= (qso != 0) * self.desi_mask.QSO_SOUTH
+        targets['OBSCONDITIONS'] |= (qso != 0)  * self.obsconditions.mask(
+            self.desi_mask.QSO.obsconditions)
+        targets['OBSCONDITIONS'] |= (qso != 0)  * self.obsconditions.mask(
+            self.desi_mask.QSO_SOUTH.obsconditions)
+
+        # Select QSO contaminants for ELG targets.  There'd be no morphology cut
+        # and we're missing the WISE colors, so the density won't be right.
+        elg = isELG(gflux=gflux, rflux=rflux, zflux=zflux)
+        targets['DESI_TARGET'] |= (elg != 0) * self.desi_mask.ELG
+        targets['DESI_TARGET'] |= (elg != 0) * self.desi_mask.ELG_SOUTH
+        targets['OBSCONDITIONS'] |= (elg != 0)  * self.obsconditions.mask(
+            self.desi_mask.ELG.obsconditions)
+        targets['OBSCONDITIONS'] |= (elg != 0)  * self.obsconditions.mask(
+            self.desi_mask.ELG_SOUTH.obsconditions)
+
+        truth['CONTAM_TARGET'] |= (elg != 0) * self.contam_mask.ELG_IS_QSO
+        truth['CONTAM_TARGET'] |= (elg != 0) * self.contam_mask.ELG_CONTAM
+
+    def sky_select(self, targets, truth, boss_std=None):
+        """Select SKY targets."""
+
+        targets['DESI_TARGET'] |= self.desi_mask.mask('SKY')
+        targets['OBSCONDITIONS'] |= self.obsconditions.mask(
+            self.desi_mask.SKY.obsconditions)
+
+    def density_select(self, targets, truth, source_name, target_name, density=None, subset=None):
+        """Downsample a target sample to a desired number density in targets/deg2."""
+        nobj = len(targets)
+
+        unique_bricks = list(set(targets['BRICKNAME']))
+        n_brick = len(unique_bricks)
+
+        for thisbrick in unique_bricks:
+            brickindx = np.where(self.brick_info['BRICKNAME'] == thisbrick)[0]
+            brick_area = float(self.brick_info['AREA'][brickindx][0])
+
+            if subset is None:
+                onbrick = np.where( (targets['BRICKNAME'] == thisbrick) )[0]
+            else:
+                onbrick = np.where( (targets['BRICKNAME'] == thisbrick) * subset )[0]
+            #onbrick = np.where((targets['BRICKNAME'] == thisbrick) * (truth['CONTAM_TARGET'] == 0))[0]
+
+            n_in_brick = len(onbrick)
+            if n_in_brick == 0:
+                self.log.info('No {}s on brick {}.'.format(target_name, thisbrick))
+                
+            if density and n_in_brick > 0:
+                mock_density = n_in_brick / brick_area
+                desired_density = float(self.brick_info['FLUC_EBV_{}'.format(source_name)][brickindx] * density)
+                if desired_density < 0.0:
+                    desired_density = 0.0
+
+                frac_keep = desired_density / mock_density
+                if frac_keep > 1.0:
+                    self.log.warning('Density {:.0f}/deg2 (N={:g}) lower than desired {:.0f}/deg2 on brick {}.'.format(
+                        mock_density, n_in_brick, desired_density, thisbrick))
+                else:
+                    frac_toss = 1.0 - frac_keep
+                    ntoss = int( np.ceil( n_in_brick * frac_toss ) )
+
+                    self.log.info('Downsampling from {:.0f} to {:.0f} targets/deg2 (N={:g} to N={:g}) on brick {}.'.format(
+                        mock_density, desired_density, n_in_brick, n_in_brick - ntoss, thisbrick))
+                    
+                    toss = self.rand.choice(onbrick, ntoss, replace=False)
+                    targets['DESI_TARGET'][toss] = 0
+
+    def contaminants_select(self, targets, truth, source_name, target_name, contam):
+        """Downsample contaminants to a desired number density in targets/deg2."""
+        nobj = len(targets)
+
+        unique_bricks = list(set(targets['BRICKNAME']))
+        n_brick = len(unique_bricks)
+
+        for thisbrick in unique_bricks:
+            brickindx = np.where(self.brick_info['BRICKNAME'] == thisbrick)[0]
+            brick_area = float(self.brick_info['AREA'][brickindx][0])
+
+            for contam_name in contam.keys():
+                onbrick = np.where(
+                    (targets['BRICKNAME'] == thisbrick) *
+                    #(targets['DESI_TARGET'] == 0) *
+                    (truth['CONTAM_TARGET'] & self.contam_mask.mask('{}_IS_{}'.format(target_name, contam_name)) != 0)
+                    )[0]
+                n_in_brick = len(onbrick)
+
+                if n_in_brick == 0:
+                    self.log.debug('No {}_IS_{} contaminants on brick {}.'.format(target_name, contam_name, thisbrick))
+                
+                if n_in_brick > 0:
+                    contam_density = len(onbrick) / brick_area
+                    desired_density = float(self.brick_info['FLUC_EBV_{}'.format(source_name)][brickindx] * contam[contam_name])
+                    if desired_density < 0.0:
+                        desired_density = 0.0
+
+                    frac_keep = desired_density / contam_density
+                    if frac_keep > 1.0:
+                        self.log.warning('Density {:.0f}/deg2 (N={:g}) of {} contaminants lower than desired {:.0f}/deg2 on brick {}.'.format(
+                            contam_density, n_in_brick, contam_name.upper(), desired_density, thisbrick))
+
+                    else:
+                        frac_toss = 1.0 - frac_keep
+                        ntoss = int( np.ceil( n_in_brick * frac_toss ) )
+
+                        self.log.info('Downsampling {}_IS_{} contaminants from {:.0f} to {:.0f} targets/deg2 (N={:g} to N={:g}) on brick {}.'.format(
+                            target_name, contam_name, contam_density, desired_density, n_in_brick, n_in_brick - ntoss, thisbrick))
+
+                        # This isn't quite right because we occassionally throw
+                        # away too many other "real" targets.
+                        toss = self.rand.choice(onbrick, ntoss, replace=False)
+                        targets['DESI_TARGET'][toss] = 0
